pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dicom-classifier-demo"
version = "0.1.0"
description = "Medical Imaging AI Demo — DICOM CT classification with ROI detection"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydicom>=2.4",
    "torch>=2.1",
    "torchvision>=0.16",
    "Pillow>=10.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1",
    "pydantic-settings>=2.1",
    "structlog>=24.1",
    "numpy>=1.26",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # Allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
